requirements.txt
click==7.1.2
Flask==1.1.2
Flask-BasicAuth==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-dotenv==0.14.0
Werkzeug==1.0.1
